@@ -1,6 +1,6 @@
 {
   "name": "@xyo-network/tool-dapploy-nodejs",
-  "version": "0.1.3",
+  "version": "0.1.4",
   "license": "MIT",
   "description": "Dapploy is a handy CLI tool for dApps. Create, build and dapploy your Smart Contracts!",
   "repository": {
@@ -11,18 +11,18 @@
     "url": "https://github.com/XYOracleNetwork/tool-dapploy-nodejs/issues"
   },
   "dependencies": {
-    "aws-sdk": "^2.316.0",
+    "aws-sdk": "^2.382.0",
     "commander": "^2.16.0",
     "configparser": "^0.2.6",
     "ipfs-api": "^26.1.2",
-    "shelljs": "^0.8.2",
-    "truffle": "^4.1.13",
+    "shelljs": "^0.8.3",
+    "truffle": "^5.0.0",
     "untildify": "^3.0.3",
-    "web3": "^1.0.0-beta.35"
+    "web3": "^1.0.0-beta.37"
   },
   "devDependencies": {
     "babel-eslint": "^10.0.1",
-    "eslint": "^5.8.0",
+    "eslint": "^5.11.1",
     "eslint-config-airbnb": "^17.1.0",
     "eslint-config-standard": "^12.0.0",
     "eslint-config-standard-react": "^7.0.2",
@@ -30,9 +30,9 @@
     "eslint-plugin-jsx-a11y": "^6.1.1",
     "eslint-plugin-node": "^8.0.0",
     "eslint-plugin-promise": "^4.0.0",
-    "eslint-plugin-react": "^7.11.1",
+    "eslint-plugin-react": "^7.12.1",
     "eslint-plugin-standard": "^4.0.0",
-    "ganache-cli": "^6.1.7"
+    "ganache-cli": "^6.2.5"
   },
   "publishConfig": {
     "registry": "https://registry.npmjs.org/"
